refactor(modal): lift the web focus trap and the touch slop out of the chrome
useWebModalFocus, the dialog and stack-entry accessibility props and the
dialog test ID move to webModal/modal-accessibility.ts, and TOUCH_SLOP moves
to constants/gestures.ts. Both are a straight move: the React Native chrome
imports them and behaves exactly as before.

They come out because a second chrome is about to need the same code, and a
focus trap and an accessibility contract are the last things worth having two
copies of.
